Rename ows_util to image_util
